Add files via upload
smoothing:
Smoothing (or blurring) in image processing is a technique used to reduce noise and suppress unnecessary details in an image while preserving important structures. It is commonly used in preprocessing before edge detection, segmentation, or other analysis.
sharpening:
Image sharpening in image processing is often achieved by applying a filter that enhances the edges and fine details in an image. The most commonly used method is convolution with a sharpening kernel.